add http methods to the packet generator
